Use common database pattern for kick and current setpoint sync.
Previously, we depended on the value of Current-SP eventually settling,
because the direct and inverse operations in use should converge onto a
specific Kick and Current pair. However, any update of Current-SP would
cause at least one write operation into Current-SP due to processing of
Kick-SP, including the update caused by a value change of Kick-SP. In
normal operating conditions, this is not a big issue, but, for this IOC,
when we don't have an electron beam and the dipoles are turned off (0
GeV beam energy), all current values are converted into 0 urad kicks,
which are then converted into 0 A currents. This makes it impossible to
operate the fast corrector power supplies, since this IOC is constantly
forcing the Current-SP values to 0 whenever a different value is written
into them.

Now, since we are using SDIS and PACT, we can specifically disable any
writes into Current-SP when updating Kick-SP based on it, but writes
into Kick-SP still update Current-SP accordingly.

[1] https://docs.epics-controls.org/en/latest/process-database/common-database-patterns.html#readback-from-a-different-hardware-address